+++ ChangeLog.md
@@ -3,5 +3,7 @@
 - 2020-02-11: New reps, some exercise changes. Realized I should think about
               exercise change outside of doing reps, 15 minutes.
 - 2020-02-12: New reps, 10 minutes
+- 2020-02-13: No reps :(
+- 2020-02-14: Reps, used foldMap instead of foldr for foldable, 8 minutes